+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - /about endpoint with build information
 - Version of the webapp in the webapp's header
 - ForwarderVersion and IndexerVersion in the indexed fields
+- Metric for the size of the received coredumps, and options to control the exposed buckets
+- Coredump.AnalyzedAt field with the date the coredump was anayzed (updated upon analysis)
+- Coredump.Executable field with the name of the executable
 ### Changed
 - The `c.analyzer` and `go.analyzer` options now take the gdb or delve commands to execute on the coredump, instead of full shell commands (#27)
 - Design & Interface overhaul:
@@ -18,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	- Design adjustments for better compliance with [WCAG 2.0](https://www.w3.org/TR/WCAG20/)
 - The sort and sort order parameters are now separate options because it's simpler
 - Limit the sort options to date and hostname because the executable one isn't working as intended anymore
+- Compute the core and executable sizes during indexation instead of analysis
+- Rename Coredump.Date to Coredump.DumpedAt for consistency with Coredump.AnalyzedAt
 ### Fixed
 - Case of the fields for sorting is back to lowercase (#26)